Create THE CHALLENGE: The Galactic Prime Pattern Machine
Reads a number N from the user. This represents the size of a square grid (NxN).

Fills the grid with the first N*N prime numbers.

Converts each prime number in the grid into a "symbol", based on this decision tree:

If the number is a palindrome, use @

Else if the number's digital root is 7, use #

Else if the number is divisible by 3 but not by 9, use &

Else use *

Displays the final grid of symbols, formatted like a square.